requirements.txt
google-generativeai>=0.3.0
langfuse>=2.0.0
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
